fix: don't raise error when field with default not in config
Closes #2. Fields with simple defaults (like `clipboard_timeout`) are
simply assigned a default in the dataclass. Ones with more complicated
defaults are initially passed as `None`, and then assigned a sensible
default in the `__post_init__` method of the dataclass. Also moves
`is_secret_encrypted` to the `LockeyConfig` class and uses the unix
`file` command to get the actual mime type of the file rather than
simply reading from the config file.